+++ NEWS.md
@@ -3,6 +3,7 @@
 ## bug fixes
 
 `plot<SpatVector>` used the wrong main label in some cases [#1210](https://github.com/rspatial/terra/issues/1210) by Márcia Barbosa
+`plotRGB` failed with an "ext=" argument [#1228](https://github.com/rspatial/terra/issues/1228) by Dave Edge
 
 
 ## enhancements