+++ NEWS.md
@@ -1,9 +1,10 @@
-# version 1.7-41
+# version 1.7-43
 
 ## bug fixes
 
 `plot<SpatVector>` used the wrong main label in some cases [#1210](https://github.com/rspatial/terra/issues/1210) by Márcia Barbosa
 `plotRGB` failed with an "ext=" argument [#1228](https://github.com/rspatial/terra/issues/1228) by Dave Edge
+`rast<array>` failed badly when the array had less than three dimensions. [#1254](https://github.com/rspatial/terra/issues/1254) by andreimirt.
 
 
 ## enhancements